Put classes in package duke
Let's organize classes into a suitable package.
