+++ src/main/java/duke/Event.java
@@ -1,3 +1,5 @@
+package duke;
+
 /**
  * Represents an event with start and end time.
  */
