+++ src/main/java/duke/ListCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ListCommand extends Command {
     public ListCommand() {
